requirements.txt
pygame-ce==2.5.2
pygbag==0.9.2